requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
rembg==2.0.57
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.4
onnxruntime==1.17.1